Add <slot> to each square
To allow custom SVGs to be rendered onto squares, add a named `<slot>` to each
square, with name corresponding to the label of the square.

According to the HTML spec for slot name, changing the name of a slot
causes re-assignment of slots (https://dom.spec.whatwg.org/#slot-name),
so this should automatically adjust when flipping the orientation of the
board.