fix: Fix mobile drag bug (#22)
Fixes #20.

Code was arranged so that move "ending" logic would be happening in the
click handler only. However, in some cases (on mobile) a click event
won't be fired, mostly after a drag is started.

To work around this, this code duplicates click handling logic to
`pointerup`, and adds a flag that prevents re-handling that event in
the `click` handler.