Disable browser touch actions on moveable squares (#54)
Earlier, the whole board had `touch-action: none` set. This was too
aggressive, and prevented scroll when the board covered the whole width
of the screen. This was pointed out in #51.

This commit moves the `touch-action: none` to squares with `.moveable`
class, as those are the only ones that would support a drag action
(where we wouldn't want the browser interfering).